Move condition and loinc controllers
